Move functions to pytorch_lightning.utilities.data
